Abstract Session and Day Models
Abstracted Session and Day models to separate files
for easier coordination with the backend.
