+++ lib/models/session.dart
@@ -1,0 +1,11 @@
+class Session {
+  String time;
+  List<String> students;
+  int openSlots;
+
+  Session(String time, List<String> students, int openSlots) {
+    this.time = time;
+    this.students = students;
+    this.openSlots = openSlots;
+  }
+}
